+++ backend/stactools-hotosm/src/stactools/hotosm/stac.py
@@ -169,6 +169,13 @@ def create_item(oam_metadata: OamMetadata) -> Item:
     if oam_metadata.uploaded_at:
         item.properties["created"] = datetime_to_str(_as_utc(oam_metadata.uploaded_at))
 
+    if oam_metadata.uploader_id:
+        item.properties["oam:uploader_id"] = oam_metadata.uploader_id
+    if oam_metadata.uploader_name:
+        item.properties["oam:uploader_name"] = oam_metadata.uploader_name
+    if oam_metadata.uploader_email:
+        item.properties["oam:uploader_email"] = oam_metadata.uploader_email
+
     item.add_asset(
         "visual",
         Asset(