+++ backend/uploader-api/pipeline/metadata/metadata.py
@@ -644,6 +644,9 @@ def build_item(
         image_file_size=os.path.getsize(cog_path),
         thumbnail_url=thumbnail_url,
         metadata_url=metadata_url,
+        uploader_id=user_md.get("uploader_id"),
+        uploader_name=user_md.get("uploader_name"),
+        uploader_email=user_md.get("uploader_email"),
     ).sanitize()
 
     item = create_item(oam)